Remove 'uglify' and 'skip' strategies: always wrap 1st positional arg in `params`
The only key in a Rails 4.2 app that might actually deserve to be
outside of `params` is :format.

Internally, when you provide `format` as a keyword argument to
`get`, Rails just adds it to the params hash anyway. So it doesn't
really matter if it's inside or outside.

`rails5-spec-converter` will still try to bring it outside
if the invocation is using a hash literal.